refactor: 상품 등록요청 관련 패키지 이동
