+++ backend/src/main/java/com/team5/backend/domain/member/admin/dto/ProductRequestUpdateReqDto.java
@@ -1,6 +1,6 @@
 package com.team5.backend.domain.member.admin.dto;
 
-import com.team5.backend.domain.member.admin.entity.ProductRequestStatus;
+import com.team5.backend.domain.member.productrequest.entity.ProductRequestStatus;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
