+++ backend/src/main/java/com/team5/backend/domain/member/productrequest/dto/ProductRequestCreateReqDto.java
@@ -1,4 +1,4 @@
-package com.team5.backend.domain.member.admin.dto;
+package com.team5.backend.domain.member.productrequest.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
